Update VueTagInput.vue
Deleting the last tag on backspace is very mistake prone for the user. Often a user will hit backspace a bunch of times to clear the input and may accidentally delete several of their tags. Can we add deleteOnBackspace prop to allow turning off "delete tag if key backspace at the start of input"?